Resolve conflict in server/server/src/lib.rs:
- Use develop's refactored startup sequence (server starts before migrations,
  with MigratingDatabase -> Initialising -> Operational status progression)
- Incorporate RC's configurable actix-web worker count (#10269)

Resolve conflict in package.json:
- Keep develop's version (2.20.00)

+++ server/service/src/preference/preferences/number_of_months_threshold_to_show_over_stock_alerts_for_products.rs
@@ -3,7 +3,7 @@ use crate::preference::{PrefKey, Preference, PreferenceType, PreferenceValueType
 pub struct NumberOfMonthsThresholdToShowOverStockAlertsForProducts;
 
 impl Preference for NumberOfMonthsThresholdToShowOverStockAlertsForProducts {
-    type Value = i32;
+    type Value = f64;
 
     fn key(&self) -> PrefKey {
         PrefKey::NumberOfMonthsThresholdToShowOverStockAlertsForProducts
@@ -14,6 +14,6 @@ impl Preference for NumberOfMonthsThresholdToShowOverStockAlertsForProducts {
     }
 
     fn value_type(&self) -> PreferenceValueType {
-        PreferenceValueType::Integer
+        PreferenceValueType::Float
     }
 }
